dhcpv4: combine small files
