+++ dhcpv4/option_misc.go
@@ -1,6 +1,7 @@
 package dhcpv4
 
 import (
+	"github.com/insomniacslk/dhcp/iana"
 	"github.com/insomniacslk/dhcp/rfc1035label"
 )
 
@@ -9,4 +10,9 @@ import (
 // The domain search option is described by RFC 3397, Section 2.
 func OptDomainSearch(labels *rfc1035label.Labels) Option {
 	return Option{Code: OptionDNSDomainSearchList, Value: labels}
+}
+
+// OptClientArch returns a new Client System Architecture Type option.
+func OptClientArch(archs ...iana.Arch) Option {
+	return Option{Code: OptionClientSystemArchitectureType, Value: iana.Archs(archs)}
 }
